feat: add game mode selector
Currently, only Guess Who > Years works, as it was before
More game modes to come...
Two layout components were also abstracted. More to come...
