+++ src/app/guess/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+  return (
+    <div className="flex items-center justify-center h-screen">
+      <h1 className="text-4xl font-bold">In Construction</h1>
+    </div>
+  );
+}
